Remove un-called code after Squeel removal
The only place this was used was in a now-eliminated Squeel query for
retrieving records with a matching value for a given key, like this:

    where{cast(Squeel::Nodes::Stub.new(store_attribute).op('->', name.to_s).as(sql_type)) == val}

But now that we're storing in a JSONB column, the types should be
preserved and this should be unnecessary. If this commit is still here,
then the specs must have passed without needing the type casting.